+++ cx-portal/src/features/serviceManagement/apiSlice.ts
@@ -28,6 +28,11 @@ export enum ReleaseProcessTypes {
   SERVICE_RELEASE = 'serviceRelease',
 }
 
+export enum ServiceTypeIdsEnum {
+  CONSULTANCE_SERVICE = 'CONSULTANCE_SERVICE',
+  DATASPACE_SERVICE = 'DATASPACE_SERVICE',
+}
+
 export type CreateServiceStep1Item = {
   title?: string
   price?: string | null
@@ -180,6 +185,15 @@ export const apiSlice = createApi({
         method: 'PUT',
       }),
     }),
+    fetchDocument: builder.mutation({
+      query: (obj) => ({
+        url: `/api/services/${obj.serviceId}/serviceDocuments/${obj.documentId}`,
+        responseHandler: async (response) => ({
+          headers: response.headers,
+          data: await response.blob(),
+        }),
+      }),
+    }),
   }),
 })
 
@@ -195,4 +209,5 @@ export const {
   useFetchNewDocumentByIdMutation,
   useSubmitServiceMutation,
   useFetchFrameDocumentByIdMutation,
+  useFetchDocumentMutation,
 } = apiSlice